update environment.py
added a method to get the reference file into a numpy array in order to do the error calculation, started to make the necessary method to get the spectra from the created particle. For the moment it has to go through writing a file even if it won't be necessary on the long run.